Minor bug fixes, Groovy version bump
